+++ src/server/core/config.ts
@@ -11,7 +11,7 @@ import {
   pathToFileUrl,
   resolve,
 } from "@dreamer/runtime-adapter";
-import { $tr } from "../utils/i18n.ts";
+import { $tr } from "../../i18n.ts";
 import { logger } from "../utils/logger.ts";
 import type {
   AppBuildConfig,